backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.19
numpy==2.2.0
pillow==11.0.0
trimesh==4.5.3
scipy==1.14.1
blake3==1.0.4
psutil==6.1.0
redis==5.2.1
tenacity==9.0.0
prometheus-client==0.21.1
